+++ src/index.html
@@ -7,12 +7,9 @@
 	</head>
 	<body>
 		<h1>Test</h1>
-
-		<script src="../src/js/API/gmaps.js"></script>
-		<script src="../src/js/API/zomato.js"></script>
+		<div id="restaurants"></div>
+		<!-- <script src="../src/js/API/gmaps.js"></script>
+		<script src="../src/js/API/zomato.js"></script> -->
 		<script src="../src/js/test.js"></script>
 	</body>
 </html>
-
-<!-- shaf was here -->
-<!-- shaf was here part 2 -->
