+++ src/js/API/gmaps.js
@@ -1,15 +1,15 @@
-// importing zomato.js
-// export to index.js
+// // importing zomato.js
+// // export to index.js
 
-// Create the script tag, set the appropriate attributes
-var script = document.createElement('script');
-script.src = 'https://maps.googleapis.com/maps/api/js?key=YOUR_API_KEY&callback=initMap';
-script.defer = true;
+// // Create the script tag, set the appropriate attributes
+// var script = document.createElement('script');
+// script.src = 'https://maps.googleapis.com/maps/api/js?key=YOUR_API_KEY&callback=initMap';
+// script.defer = true;
 
-// Attach your callback function to the `window` object
-window.initMap = function () {
-    // JS API is loaded and available
-};
+// // Attach your callback function to the `window` object
+// window.initMap = function () {
+//     // JS API is loaded and available
+// };
 
-// Append the 'script' element to 'head'
-document.head.appendChild(script);
+// // Append the 'script' element to 'head'
+// document.head.appendChild(script);
